Got Autograd working for multidimensional tanh!
